+++ frontend/src/components/base/badge/index.ts
@@ -1,0 +1,1 @@
+export * from "./base";
